+++ test/components/folio/console/tiptap/overlay/form_component_test.rb
@@ -23,6 +23,43 @@ class Folio::Console::Tiptap::Overlay::FormComponentTest < Folio::Console::Compo
     }
   end
 
+  class CarouselItem < Folio::Tiptap::Node
+    tiptap_node nested: true,
+                structure: {
+                  url: { type: :url_json, disable_label: true },
+                  cover: :image,
+                  badge_label: :string,
+                  title: :string,
+                }
+  end
+
+  class Carousel < Folio::Tiptap::Node
+    tiptap_node structure: {
+      items: {
+        type: :nested_nodes,
+        node_class: CarouselItem,
+      },
+    }
+  end
+
+  class FlatCarouselItem < Folio::Tiptap::Node
+    tiptap_node nested: true,
+                structure: {
+                  cover: :image,
+                  title: :string,
+                },
+                form_layout: nil
+  end
+
+  class FlatCarousel < Folio::Tiptap::Node
+    tiptap_node structure: {
+      items: {
+        type: :nested_nodes,
+        node_class: FlatCarouselItem,
+      },
+    }
+  end
+
   def test_render
     with_controller_class(Folio::Console::PagesController) do
       with_request_url "/console/pages/new" do
@@ -82,6 +119,31 @@ class Folio::Console::Tiptap::Overlay::FormComponentTest < Folio::Console::Compo
         render_inline(Folio::Console::Tiptap::Overlay::FormComponent.new(node:))
 
         assert_selector('[name="tiptap_node_attrs[data][cards][item_0][data][title]"].is-invalid')
+      end
+    end
+  end
+
+  test "renders default aside attachment layout for nested node rows" do
+    with_controller_class(Folio::Console::PagesController) do
+      with_request_url "/console/pages/new" do
+        render_inline(Folio::Console::Tiptap::Overlay::FormComponent.new(node: Carousel.new))
+
+        assert_selector(".f-c-tiptap-overlay-form-nested-nodes__item-fields .f-c-tiptap-overlay-form-layout__row:nth-of-type(2) > .f-c-tiptap-overlay-form-layout__col:first-child [name='tiptap_node_attrs[data][items][item_0][data][cover_placement_attributes][file_id]']",
+                        visible: :all)
+        assert_selector(".f-c-tiptap-overlay-form-nested-nodes__item-fields .f-c-tiptap-overlay-form-layout__row:nth-of-type(2) > .f-c-tiptap-overlay-form-layout__col:nth-child(2) [name='tiptap_node_attrs[data][items][item_0][data][badge_label]']")
+        assert_selector(".f-c-tiptap-overlay-form-nested-nodes__item-fields .f-c-tiptap-overlay-form-layout__row:nth-of-type(2) > .f-c-tiptap-overlay-form-layout__col:nth-child(2) [name='tiptap_node_attrs[data][items][item_0][data][title]']")
+      end
+    end
+  end
+
+  test "renders flat nested node rows with explicit nil form_layout" do
+    with_controller_class(Folio::Console::PagesController) do
+      with_request_url "/console/pages/new" do
+        render_inline(Folio::Console::Tiptap::Overlay::FormComponent.new(node: FlatCarousel.new))
+
+        assert_selector("[name='tiptap_node_attrs[data][items][item_0][data][cover_placement_attributes][file_id]']",
+                        visible: :all)
+        assert_selector("[name='tiptap_node_attrs[data][items][item_0][data][title]']")
       end
     end
   end